refactor(provider): split CoordinatorClient god-file by concern (1173->213 LOC)
Extract self-contained top-level types into their own files (CoordinatorClient
Types: events/config/messages/errors; State: atomic stats + provider state +
concurrency primitives; OutboundRouter; ReachabilityMonitor) and split the
actor into +Connection (reconnect/session/receive + telemetry), +Inbound
(frame->event dispatch), +Registration (register + heartbeat), and +Outbound
(wire encoding).

Bodies byte-identical; 21 members reached across the split widened
private->internal. The file-private `Logger` typealias was renamed to a unique
`CoordinatorWSLogger` (kept internal) so widening the actor's logger property
doesn't shadow `Logging.Logger`/`os.Logger` elsewhere in the module.

Build green; swift test 1052/73 pass (unchanged from baseline).
